+++ src/vue/components/ImageLoader.vue
@@ -8,9 +8,9 @@
       <p><i class="fa fa-times fa-fw fa-2x"></i> {{ error }}</p>
       <button @click="loadImage" class="reload-image-btn">Reload Image</button>
     </div>
-    <div class="qcw-image-container" v-if="!isImage && !isLoading">
+    <div class="qcw-file-container" v-if="!isImage && !isLoading">
       <a :href="uri" target="_blank">
-        <i class="fa fa-file-o" aria-hidden="true"><div class="ft">File</div><div class="fe">{{ ext }}</div></i>
+        <i class="fa" aria-hidden="true" :class="fileClassName"></i>
         <div class="file-name">{{ filename }}</div>
       </a>
     </div>
@@ -32,6 +32,18 @@
         error: ''
       }
     },
+    computed: {
+      fileClassName() {
+        const ext = this.ext.toLowerCase()
+        const videos = ['mov','mp4','avi','mkv']
+        const images = ['jpg','jpeg','gif','webp','png']
+        const archives = ['tar', 'zip', 'rar', 'iso']
+        if ( videos.indexOf(ext) > -1 ) return 'fa-file-video-o';
+        if ( images.indexOf(ext) > -1 ) return 'fa-file-image-o';
+        if ( archives.indexOf(ext) > -1 ) return 'fa-file-archive-o';
+        return 'fa-file-text'
+      }
+    },
     created() {
       this.loadImage();
     },
@@ -51,7 +63,7 @@
           if (this.readyState == 4 && this.status == 200){
             self.imageSrc  = URL.createObjectURL(this.response);
             self.isLoading = false;
-            setTimeout( () => self.callback(), 0 );
+            // setTimeout( () => self.callback(), 0 );
           } 
         }
         xhr.open('GET', self.uri, true);
@@ -69,6 +81,7 @@
 </script>
 
 <style lang="scss">
+  .image-container { height: 100px; }
   .reload-image-btn {
     display: block;
     width: 100%;