+++ src/vue/main.js
@@ -1,9 +1,14 @@
 import QiscusSDK from '../sdk'
 import Vue from 'vue'
 import QiscusWidget from './App'
+import Toasted from 'vue-toasted';
+
+Vue.use(Toasted, {
+  position: 'bottom-right',
+  duration: 2000
+})
 
 // QiscusSDK.mode = 'wide';
-// QiscusSDK.setUser('fikri@qiscus.com', 'password', 'fikri');
 // QiscusSDK.init({
 //   AppId: 'dragongo',
 //   options: {
@@ -12,6 +17,7 @@ import QiscusWidget from './App'
 //     }
 //   }
 // });
+// QiscusSDK.setUser('fikri@qiscus.com', 'password', 'fikri');
 
 /* let's render the view layer */
 window.vm = new Vue({