+++ src/vue/vuex/actions.js
@@ -1,10 +1,11 @@
 import find from 'lodash/fp/find'
 
 export default {
+  toggleDevMode: ({commit, state}) => commit('TOGGLE_DEV_MODE'),
   changeRoom: ({commit, state}, room) => commit('CHANGE_ROOM', room),
   toggleChatWindow: ({commit, state}) => commit('TOGGLE_CHAT_WINDOW'),
   chatTarget: ({commit}, {email, options = {}}) => {
-    qiscus.chatTarget(email, options)
+    return qiscus.chatTarget(email, options)
     .then((response) => {
       commit('CHAT_TARGET', {email, options})
       const selected = qiscus.selected.comments
@@ -17,6 +18,10 @@ export default {
         //on entering the room, wait for data processed then focus on comment form
         document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
       }, 0)
+      return Promise.resolve(qiscus.selected)
+    }, (error) => {
+      vm.$toasted.error('Error getting chat room. Please make sure the target is valid')
+      return Promise.reject(error)
     })
   },
   chatGroup: ({commit}, {id, oldSelected}) => {
@@ -37,7 +42,40 @@ export default {
     return qiscus.submitComment(topic_id, comment)
     .then((response) => {
       commit('SUBMIT_COMMENT', qiscus.selected)
+      const selected = qiscus.selected.comments
+      const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
+      setTimeout(function(){
+        if(latestCommentId > 0){
+          const elementToScroll = document.getElementById(latestCommentId)
+          elementToScroll.scrollIntoView({block: 'end', behavior: 'smooth'})
+        }
+        //on entering the room, wait for data processed then focus on comment form
+        document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
+      }, 0)
       return Promise.resolve(qiscus.selected);
+    }, (error) => {
+      return Promise.reject(error)
+    })
+  },
+  resendComment: ({commit}, comment) => {
+    return qiscus.resendComment(comment)
+    .then((response) => {
+      commit('SUBMIT_COMMENT', qiscus.selected)
+      const selected = qiscus.selected.comments
+      const latestCommentId = (selected.length > 0) ? selected[selected.length-1].id : 0
+      setTimeout(function(){
+        if(latestCommentId > 0){
+          const elementToScroll = document.getElementById(latestCommentId)
+          elementToScroll.scrollIntoView({block: 'end', behavior: 'smooth'})
+        }
+        //on entering the room, wait for data processed then focus on comment form
+        document.getElementsByClassName('qcw-comment-form').item(0).getElementsByTagName('textarea').item(0).focus();
+      }, 0)
+      console.info('resend comment successful')
+      return Promise.resolve(qiscus.selected);
+    }, error => {
+      console.error('resend comment error', error)
+      return Promise.reject(error)
     })
   },
   loadComments: ({commit}, payload) => commit('LOAD_COMMENTS', payload),