+++ src/vue/vuex/store.js
@@ -14,6 +14,7 @@ const state = {
   selected: qiscus.selected,
   windowStatus: false,
   participants: qiscus.participants,
+  plugins: qiscus.plugins,
   // mqtt: new MqttAdapter("wss://mqtt.qiscus.com:1886", callbacks),
   mqtt: new MqttAdapter("wss://mqtt.qiscus.com:1886/mqtt", MqttCallback),
   mqttData: {
@@ -22,11 +23,15 @@ const state = {
   init: qiscus.isInit,
   isLoadingComments: false,
   imageModalLink: '',
-  imageModalOn: false
+  imageModalOn: false,
+  dev_mode: false
 }
 
 // Create an object storing various mutations. We will write the mutation
 const mutations = {
+  TOGGLE_DEV_MODE (state) {
+    state.dev_mode = !state.dev_mode
+  },
   CHANGE_ROOM (state, room) {
     state.selected = room
   },
@@ -40,6 +45,7 @@ const mutations = {
     state.selected = state.qiscus.selected;
     state.mqtt.subscribe(`r/${state.selected.id}/${state.selected.last_comment_topic_id}/+/t`);
     state.mqtt.subscribe(`${state.qiscus.userData.token}/c`);
+    state.mqttData.typing = '';
   },
   CHAT_GROUP (state, {id, oldSelected}) {
     if(state.selected) {
@@ -49,6 +55,7 @@ const mutations = {
     }
     state.windowStatus = true;
     state.selected = state.qiscus.selected;
+    state.mqttData.typing = '';
     state.mqtt.subscribe(`r/${state.selected.id}/${state.selected.last_comment_topic_id}/+/t`);
     state.mqtt.subscribe(`${state.qiscus.userData.token}/c`);
   },
@@ -78,10 +85,10 @@ const mutations = {
     // })
   },
   SET_TYPING (state, payload) {
-    if(payload.topic == state.qiscus.email) return
+    if(payload.topic.username == state.qiscus.email || payload.topic.room_id != state.selected.id) return
     // let's get the email of this payload
-    const Participant = state.qiscus.selected.participants.find( (participant) => participant.email == payload.topic )
-    const username = (Participant) ? Participant.username : payload.topic 
+    const Participant = state.qiscus.selected.participants.find( (participant) => participant.email == payload.topic.username )
+    const username = (Participant) ? Participant.username : payload.topic.username 
 
     if(payload.message == 1){
       state.mqttData.typing = username;